requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-gitlab==4.3.0
litellm==1.14.0
openai==1.5.0
tiktoken>=0.8.0
python-multipart==0.0.6
aiohttp==3.9.1
asyncio-throttle==1.0.2
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
jinja2==3.1.2
pyyaml==6.0.1
tenacity==8.2.3
prometheus-client==0.19.0